requirements.txt
asgiref
Django
fake-factory
Faker
pillow
python-dateutil
sqlparse
typing_extensions
tzdata
gunicorn
langchain
python-dotenv
daphne
youtube-transcript-api
unstructured
validators==0.28.1
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
pandas
langchain-groq
pytube
numexpr
mysqlclient